Extract common test logic in FullStory destination tests
